fix: правки после ревью
