+++ packages/core/components/page-title/__stories__/screenshot/screenshot.stories.ts
@@ -8,7 +8,3 @@ export default {
 export const Screenshot: Story = {
   name: 'Заголовок с названием страницы и кнопка удаления страницы выровнены вертикально по центру',
 };
-
-export const ScreenshotTwo: Story = {
-  name: 'Скриншотег для проверки',
-};
