+++ packages/core/package.json
@@ -25,7 +25,6 @@
     "@jest/globals": "^29.7.0",
     "@colibrijs/mocks": "^1.0.0",
     "@colibrijs/wait-storybook": "^1.0.0",
-    "@colibrijs/test-utils": "^1.0.0",
     "@storybook/addon-actions": "^8.1.10",
     "@storybook/addon-essentials": "^8.1.10",
     "@storybook/addon-interactions": "^8.1.10",
