--- packages/test-utils/index.ts
@@ -1,1 +1,0 @@
-export { withScreenshotsDecorator } from './screenshots-comparator/decorator';
