--- packages/test-utils/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@colibrijs/test-utils",
-  "private": true,
-  "version": "1.0.0",
-  "description": "Пакет с утилитами для тестов",
-  "devDependencies": {
-    "@storybook/manager-api": "^8.4.6"
-  },
-  "peerDependencies": {
-    "react": "^18"
-  }
-}
